Utilsフォルダを作成しLoadingStatusのEnumをそこへ移動
